+++ app/controllers/confirmations_controller.rb
@@ -1,3 +1,4 @@
 class ConfirmationsController < ApplicationController
   def new
-    
+  end
+end
