+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20171019211908) do
+ActiveRecord::Schema.define(version: 20171021030141) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -90,6 +90,8 @@ ActiveRecord::Schema.define(version: 20171019211908) do
     t.datetime "oauth_expires_at"
     t.string "image"
     t.bigint "vendor_id"
+    t.string "verification_code"
+    t.string "phone_number"
     t.index ["vendor_id"], name: "index_users_on_vendor_id"
   end
